Added plan flag for infrastructure deploy

+++ cdflow_commands/plugins/base.py
@@ -5,13 +5,19 @@ from subprocess import check_call
 class Destroy(object):
 
     def __init__(
-        self, boto_session, component_name, environment_name, bucket_name
+        self,
+        boto_session,
+        component_name,
+        environment_name,
+        bucket_name,
+        plan_only
     ):
         self._boto_session = boto_session
         self._aws_region = boto_session.region_name
         self._component_name = component_name
         self._environment_name = environment_name
         self._bucket_name = bucket_name
+        self._plan_only = plan_only
 
     @property
     def _terraform_parameters(self):
@@ -33,14 +39,19 @@ class Destroy(object):
             ['terraform', 'plan', '-destroy'] + self._terraform_parameters,
             env=env
         )
-        check_call(
-            ['terraform', 'destroy', '-force'] + self._terraform_parameters,
-            env=env
-        )
-        boto_s3_client = self._boto_session.client('s3')
-        boto_s3_client.delete_object(
-            Bucket=self._bucket_name,
-            Key='{}/{}/terraform.tfstate'.format(
-                self._environment_name, self._component_name
+        if not self._plan_only:
+            check_call(
+                [
+                    'terraform',
+                    'destroy',
+                    '-force'
+                ] + self._terraform_parameters,
+                env=env
             )
-        )
+            boto_s3_client = self._boto_session.client('s3')
+            boto_s3_client.delete_object(
+                Bucket=self._bucket_name,
+                Key='{}/{}/terraform.tfstate'.format(
+                    self._environment_name, self._component_name
+                )
+            )
